+++ src/main/java/com/fleet/status/controller/AircraftController.java
@@ -3,9 +3,9 @@ package com.fleet.status.controller;
 import com.fleet.status.entity.Aircraft;
 import com.fleet.status.entity.Carrier;
 import com.fleet.status.entity.Type;
-import com.fleet.status.service.impl.AircraftService;
-import com.fleet.status.service.impl.CarrierService;
-import com.fleet.status.service.impl.TypeService;
+import com.fleet.status.service.AircraftService;
+import com.fleet.status.service.CarrierService;
+import com.fleet.status.service.TypeService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
