+++ src/main/java/com/fleet/status/dao/repository/AircraftRepository.java
@@ -6,13 +6,8 @@ import org.springframework.data.jpa.repository.query.Procedure;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
-
 @Profile("!test")
 public interface AircraftRepository extends CrudRepository<Aircraft, Integer> {
     @Procedure(procedureName = "uspDeleteAircraft")
     void deleteAircraft(@Param("intAircraftId") int id);
-
-    @Procedure(procedureName = "uspShowCarrierAircraft")
-    List<Aircraft> getAllAircraftFromCarrier(@Param("intCarrierId") int carrierId);
 }
