+++ src/main/java/com/fleet/status/entity/Event.java
@@ -1,17 +1,11 @@
 package com.fleet.status.entity;
 
 import com.opencsv.bean.CsvBindByName;
-import com.opencsv.bean.CsvBindByPosition;
 import com.opencsv.bean.CsvDate;
 import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.Getter;
-import lombok.ToString;
+import lombok.*;
 
-import java.time.Duration;
 import java.time.Instant;
-import java.util.Date;
 import java.util.List;
 
 @Data
@@ -59,20 +53,7 @@ public class Event {
     @Getter(AccessLevel.NONE)
     private String downTime;
 
-    public String getDownTime(){
-        if (getEndTime() == null || getStartTime() == null) {
-            return "Down time is not available";
-        }
-        Duration downtime = Duration.between(startTime, endTime);
-        return "Down Time: " + downtime.toDaysPart() + "d " + downtime.toHoursPart() + "h " + downtime.toMinutesPart() + "m";
-    }
-
     @Transient
     @CsvBindByName(column = "Tail #")
     private String csvTailNumber;
-
-    public void populateCsvFields() {
-        csvTailNumber = aircraft.getTailNumber();
-    }
-
 }
