+++ src/main/java/com/fleet/status/service/CarrierService.java
@@ -1,4 +1,4 @@
-package com.fleet.status.service.impl;
+package com.fleet.status.service;
 
 import com.fleet.status.dao.repository.CarrierRepository;
 import com.fleet.status.entity.Carrier;
@@ -14,10 +14,6 @@ import java.util.List;
 public class CarrierService {
 
     private final CarrierRepository carrierRepository;
-
-    public void save(Carrier carrier) throws Exception {
-        carrierRepository.save(carrier);
-    }
 
     public List<Carrier> getAllCarrier() {
         return (List<Carrier>) carrierRepository.findAll();
